Changed the data type of 'phi' from int to double.

+++ include/NAOS/ellipsoidPotential.hpp
@@ -74,7 +74,7 @@ void computeEllipsoidGravitationalPotential(
                                     - ( zCoordinateSquare * alphaSquare * betaSquare );
 
     // Perform a check on the sign of phi(r,0) to determine the value of lambda [3].
-    int phi = ( xCoordinateSquare / alphaSquare )
+    double phi = ( xCoordinateSquare / alphaSquare )
                  + ( yCoordinateSquare / betaSquare )
                  + ( zCoordinateSquare / gammaSquare ) - 1.0;
     Real lambda;
